move air trace data
