+++ curta/src/chip/trace/data.rs
