+++ curta/src/chip/trace/generator.rs
@@ -3,7 +3,7 @@ use alloc::sync::Arc;
 use anyhow::{Error, Result};
 use serde::{Deserialize, Serialize};
 
-use super::trae_data::AirTraceData;
+use super::data::AirTraceData;
 use super::writer::TraceWriter;
 use crate::chip::table::log_derivative::entry::LogEntry;
 use crate::chip::table::lookup::table::LookupTable;
