+++ curta/src/chip/trace/writer.rs
@@ -5,7 +5,7 @@ use std::sync::{LockResult, RwLock, RwLockReadGuard, RwLockWriteGuard};
 
 use serde::{Deserialize, Serialize};
 
-use super::trae_data::AirTraceData;
+use super::data::AirTraceData;
 use crate::chip::arithmetic::expression::ArithmeticExpression;
 use crate::chip::instruction::Instruction;
 use crate::chip::register::array::ArrayRegister;
